Refactor chat logic into hooks and fix review comments
